[SDX-1041] Create box component - removed BaseBox, using Box directly instead - minor review feedback
