+++ src/framework/base/_index.scss
@@ -1,3 +1,2 @@
 
-@import 'box/index';
 @import 'dropdown/index';
