--- src/framework/base/box/_index.scss
@@ -1,9 +1,0 @@
-
-@mixin baseBox {
-  background-color: #ffffff;
-  border: 1px solid #d4dadc;
-
-  &.roundedCorners {
-    border-radius: 5px;
-  }
-}
