+++ src/framework/body/BodyPanel.jsx
@@ -4,7 +4,7 @@ import React, {
   PropTypes,
 } from 'react';
 
-import BaseBox from '../base/box/BaseBox.jsx';
+import Box from '../box/Box.jsx';
 
 const BodyPanel = props => {
   const classes = classNames('bodyPanel', {
@@ -12,15 +12,18 @@ const BodyPanel = props => {
   });
 
   return (
-    <BaseBox classes={classes}>
-      {props.children}
-    </BaseBox>
+    <Box
+      children={props.children}
+      classes={classes}
+      roundedCorners={props.roundedCorners}
+    />
   );
 };
 
 BodyPanel.propTypes = {
-  children: PropTypes.any,
+  children: Box.propTypes.children,
   isTopFlush: PropTypes.bool,
+  roundedCorners: Box.propTypes.roundedCorners,
 };
 
 export default BodyPanel;
