+++ src/framework/body/_bodyPanel.scss
@@ -1,6 +1,5 @@
 
 .bodyPanel {
-  @include baseBox;
   margin-top: 15px;
   max-width: $maxWidth;
   padding-bottom: 15px;
