+++ src/framework/box/Box.spec.jsx
@@ -4,8 +4,18 @@ import Box from './Box.jsx';
 
 describe('Box', () => {
   describe('Props', () => {
+    describe('children', () => {
+      it('is rendered', () => {
+        const props = {
+          children: 'test',
+        };
+        const testCase = TestCaseFactory.create(Box, props);
+        expect(testCase.dom.textContent).toBe(props.children);
+      });
+    });
+
     describe('classes', () => {
-      it('are added to the element', () => {
+      it('is added to the element', () => {
         const props = {
           classes: 'testClass',
         };
@@ -13,20 +23,40 @@ describe('Box', () => {
         expect(testCase.dom.className).toContain(props.classes);
       });
 
-      it('doesn\'t replace \'box\' class', () => {
+      it('doesn\'t replace "box" class', () => {
         const props = {
           classes: 'testClass',
         };
         const testCase = TestCaseFactory.create(Box, props);
         expect(testCase.dom.className).toContain('box');
       });
+    });
 
-      it('doesn\'t add identical classes multiple times', () => {
+    describe('dataId', () => {
+      it('is added', () => {
         const props = {
-          classes: 'box box testClass',
+          dataId: 'dataId',
         };
         const testCase = TestCaseFactory.create(Box, props);
-        expect(testCase.dom.className).toBe('box testClass');
+        expect(testCase.dom.getAttribute('data-id')).toBe(props.dataId);
+      });
+    });
+
+    describe('roundedCorners', () => {
+      it('when false doesn\'t add "box--roundedCorners" class', () => {
+        const props = {
+          roundedCorners: false,
+        };
+        const testCase = TestCaseFactory.create(Box, props);
+        expect(testCase.dom.className).not.toContain('box--roundedCorners');
+      });
+
+      it('when true adds "box--roundedCorners" class', () => {
+        const props = {
+          roundedCorners: true,
+        };
+        const testCase = TestCaseFactory.create(Box, props);
+        expect(testCase.dom.className).toContain('box--roundedCorners');
       });
     });
   });
